合并重复的代码
