+++ shadowsocks-core/src/main/java/cc/springcloud/socks/network/NioLocalServer.java
@@ -32,7 +32,6 @@
 package cc.springcloud.socks.network;
 
 import cc.springcloud.socks.Constant;
-import cc.springcloud.socks.network.nio.ChangeRequest;
 import cc.springcloud.socks.network.nio.PipeWorker;
 import cc.springcloud.socks.network.nio.RemoteSocketHandler;
 import cc.springcloud.socks.network.nio.SocketHandlerBase;
@@ -41,14 +40,11 @@ import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.net.InetSocketAddress;
-import java.nio.ByteBuffer;
 import java.nio.channels.SelectionKey;
 import java.nio.channels.Selector;
 import java.nio.channels.ServerSocketChannel;
 import java.nio.channels.SocketChannel;
-import java.nio.channels.spi.SelectorProvider;
 import java.security.InvalidAlgorithmParameterException;
-import java.util.List;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
@@ -73,58 +69,28 @@ public class NioLocalServer extends SocketHandlerBase {
 
         // print server info
         logger.info("Shadowsocks-Java v{}", Constant.VERSION);
-        logger.info("Cipher: {}",config.getMethod());
-        logger.info("{} Proxy Server starts at port: {}",config.getProxyType(),config.getLocalPort());
+        logger.info("Cipher: {}", config.getMethod());
+        logger.info("{} Proxy Server starts at port: {}", config.getProxyType(), config.getLocalPort());
     }
 
     @Override
     protected Selector initSelector() throws IOException {
-        Selector socketSelector = SelectorProvider.provider().openSelector();
+        Selector socketSelector = super.initSelector();
         _serverChannel = ServerSocketChannel.open();
         _serverChannel.configureBlocking(false);
         InetSocketAddress isa = new InetSocketAddress(_config.getLocalIpAddress(), _config.getLocalPort());
         _serverChannel.socket().bind(isa);
         _serverChannel.register(socketSelector, SelectionKey.OP_ACCEPT);
-
         return socketSelector;
     }
 
-    @Override
-    protected boolean processPendingRequest(ChangeRequest request) {
-        switch (request.type) {
-            case ChangeRequest.CHANGE_SOCKET_OP:
-                SelectionKey key = request.socket.keyFor(_selector);
-                if ((key != null) && key.isValid()) {
-                    key.interestOps(request.op);
-                } else {
-                    logger.warn("NioLocalServer::processPendingRequest (drop): {} {}",key ,request.socket);
-                }
-                break;
-            case ChangeRequest.CLOSE_CHANNEL:
-                cleanUp(request.socket);
-                break;
-        }
-        return true;
+	@Override
+    protected void finishConnection(SelectionKey key) {
+        throw new UnsupportedOperationException();
     }
 
     @Override
-    protected void processSelect(SelectionKey key) {
-        // Handle event
-        try {
-            if (key.isAcceptable()) {
-                accept(key);
-            } else if (key.isReadable()) {
-                read(key);
-            } else if (key.isWritable()) {
-                write(key);
-            }
-        }
-        catch (IOException e) {
-            cleanUp((SocketChannel)key.channel());
-        }
-    }
-
-    private void accept(SelectionKey key) throws IOException {
+    protected void accept(SelectionKey key) throws IOException {
         // local socket established
         ServerSocketChannel serverSocketChannel = (ServerSocketChannel) key.channel();
         SocketChannel socketChannel = serverSocketChannel.accept();
@@ -139,78 +105,12 @@ public class NioLocalServer extends SocketHandlerBase {
         _pipes.put(socketChannel, pipe);
         _executor.execute(pipe);
     }
+	
 
-    private void read(SelectionKey key) {
-        SocketChannel socketChannel = (SocketChannel) key.channel();
-        int readCount;
-        PipeWorker pipe = _pipes.get(socketChannel);
-        byte[] data;
-
-        if (pipe == null) {
-            // should not happen
-            cleanUp(socketChannel);
-            return;
-        }
-
-        _readBuffer.clear();
-        try {
-            readCount = socketChannel.read(_readBuffer);
-        } catch (IOException e) {
-            cleanUp(socketChannel);
-            return;
-        }
-
-        if (readCount == -1) {
-            cleanUp(socketChannel);
-            return;
-        }
-
-        data = _readBuffer.array();
-        pipe.processData(data, readCount, true);
-    }
-
-    private void write(SelectionKey key) throws IOException {
-        SocketChannel socketChannel = (SocketChannel) key.channel();
-
-        List queue = (List) _pendingData.get(socketChannel);
-        if (queue != null) {
-            synchronized (queue) {
-                // Write data
-                while (!queue.isEmpty()) {
-                    ByteBuffer buf = (ByteBuffer) queue.get(0);
-                    socketChannel.write(buf);
-                    if (buf.remaining() > 0) {
-                        break;
-                    }
-                    queue.remove(0);
-                }
-
-                if (queue.isEmpty()) {
-                    key.interestOps(SelectionKey.OP_READ);
-                }
-            }
-        }
-        else {
-            logger.warn("LocalSocket::write queue = null: {}", socketChannel);
-            return;
-        }
-    }
 
     @Override
-    protected void cleanUp(SocketChannel socketChannel) {
-        //logger.warning("LocalSocket closed: " + socketChannel);
-        super.cleanUp(socketChannel);
-
-        PipeWorker pipe = _pipes.get(socketChannel);
-        if (pipe != null) {
-            pipe.close();
-            _pipes.remove(socketChannel);
-            logger.warn("LocalSocket closed: {}",pipe.socketInfo);
-        }
-        else {
-            logger.warn("LocalSocket closed (NULL): {}",socketChannel);
-        }
-
+    protected void read(SelectionKey key) {
+        super.read(key, true);
     }
 
     @Override
@@ -222,7 +122,7 @@ public class NioLocalServer extends SocketHandlerBase {
             _serverChannel.close();
             _remoteSocketHandler.close();
         } catch (IOException e) {
-            logger.warn("io exception",e);
+            logger.warn("io exception", e);
         }
         logger.info("Server closed.");
     }
